Starting with the Bootstrap folders restructure. Split bootstrap and bitrise in two different folders. Update README. Fixed reactotron-flipper missing types. Fixed some TS interfaces.
