+++ generators/app/templates/index.d.ejs
@@ -3,6 +3,7 @@ declare module '*.json';
 declare module '*.png';
 declare module 'cerealizr';
 declare module 'reactotron-apisauce';
+declare module 'reactotron-react-native/dist/flipper';
 declare module 'redux-recompose';
 <%_ if(features.login) { _%>
 declare module 'recompose';
