+++ generators/app/templates/src/app/components/AppNavigator/index.ejs
@@ -25,7 +25,7 @@ const AppNavigator = () => {
       <%_ if(features.firebaseanalytics) { _%>
       analytics().setCurrentScreen(currentRouteName, currentRouteName);
       <%_ } _%>
-    setRouteName(currentRouteName);
+      setRouteName(currentRouteName);
     }
   };
