+++ generators/bitrise/apis/bitbucketApiConfig.js
