+++ generators/bitrise/apis/githubApiConfig.js
