+++ generators/bitrise/apis/gitlabApiConfig.js
