+++ generators/bitrise/bitrise.js
@@ -1,9 +1,10 @@
 const Generator = require('yeoman-generator');
 
-const loadBitriseInfo = require('./bitriseUtils');
-const { BITRISE_YML } = require('./files');
 const bitriseInitialization = require('./tasks/bitriseInitialization');
 const nextSteps = require('./tasks/nextSteps');
+const loadBitriseInfo = require('./bitriseUtils');
+
+const BITRISE_YML = 'bitrise.yml';
 
 class BitriseInit extends Generator {
   async loadInfo() {
@@ -35,6 +36,7 @@ class BitriseInit extends Generator {
     if (this.configInfo && this.configInfo.projectName && this.configInfo.projectPath) {
       const filepathWithoutExtension = BITRISE_YML.substring(0, BITRISE_YML.lastIndexOf('.'));
       const templatePath = `${filepathWithoutExtension}.ejs`;
+
       this.fs.copyTpl(
         this.templatePath(...templatePath.split('/')),
         this.destinationPath(...this.configInfo.projectPath.concat('/bitrise.yml').split('/')),
