+++ generators/bitrise/bitriseInfo.json
