+++ generators/bitrise/bitriseUtils.js
@@ -1,5 +1,3 @@
-const fs = require('fs');
-
 const { BITRISE_PROMPTS } = require('./constants');
 
 function buildAllPrompts(context) {
@@ -47,9 +45,7 @@ function validateConfigObject(object, context) {
 async function loadBitriseInfoFile(context) {
   let configInfo = null;
   try {
-    configInfo = JSON.parse(
-      fs.readFileSync('./wolmo-bootstrap-react-native/generators/app/bitriseInfo.json').toString()
-    );
+    configInfo = require('./bitriseInfo.json');
   } catch (e) {
     console.log('The bitriseInfo.json file is written wrong'.red.underline.bold);
     configInfo = await buildAllPrompts(context);
